repatch lost https://github.com/flutter/flutter/commit/977701cf695ec106a52dde9f3659a7b132826748
