+++ packages/flutter/lib/rendering.dart
@@ -44,6 +44,7 @@ export 'src/rendering/flow.dart';
 export 'src/rendering/image.dart';
 export 'src/rendering/layer.dart';
 export 'src/rendering/list_body.dart';
+export 'src/rendering/list_wheel_viewport.dart';
 export 'src/rendering/object.dart';
 export 'src/rendering/paragraph.dart';
 export 'src/rendering/performance_overlay.dart';
